backend/pyproject.toml
[project]
name = "uv-led-platform-backend"
version = "0.1.0"
description = "Backend for the UV LED Knowledge Platform"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.2.0",
  "langchain>=0.2.0",
  "langchain-groq>=0.1.0",
  "tavily-python>=0.3.0",
  "supabase>=2.0.0",
  "apscheduler>=3.10.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.6.0",
]
